refactor: move query extends tests into __test__/extends/query/
Mirror the src layout: query-component test suites move under
src/__test__/util/extends/query/ while extendsTestClient stays at
extends/ as the shared helper for future extension-component tests.
Import-path fixes only, no behavior change.
